Refactored UpdateItemInStructure
